changed directory structure
